Add crypto module
One of the important features of Simperby is providing a light client
implementation that can be embeded in other blockchains, built into
WASM. Compiling a Rust code into WASM would require some extra constrain
ts, so the selection of cryptography-related types should be carefully
addressed. Having this abstractions allows easy replacements of the
actual implementations in the future after some thorough survey of the
contract environments of the colony chains. Note that this background is
not relevant to the Simperby project it self, because the notion of
'colony chain' does not conceptually exist here (but in the context of
PDAO beacon chain).